Refactor: separate Repository iface and impl
